+++ src/Generator.js
@@ -140,6 +140,10 @@ generate()
     let usedIvarMap = null;
     let assignedIvarMap = null;
 
+    // Addition for push/pop based block-scope-model (a model added in Raweden fork).
+    let scopeChain = [];
+    let currentScope = null;
+
     let warnings = [ ];
 
     function makeScope(node)
@@ -199,6 +203,12 @@ generate()
         return (useSelf ? "self" : "this") + "." + symbolTyper.getSymbolForClassNameAndIvarName(className, ivarName);
     }
 
+    /**
+     * Generates script that sets the initial values for instance variables.
+     * 
+     * @param  {OJClass} ojClass
+     * @return
+     */
     function generateIvarAssignments(ojClass)
     {
         let booleanIvars = [ ];
@@ -299,6 +309,14 @@ generate()
         }
     }
 
+    /**
+     * Handles NilScript Message Expression 
+     *
+     * @todo Optimize the case where the returned value is not used. It currently inserts a temporary variable anyways.
+     * 
+     * @param  {[type]} node [description]
+     * @return {[type]}      [description]
+     */
     function handleOJMessageExpression(node)
     {
         let receiver     = node.receiver.value;
@@ -742,10 +760,650 @@ generate()
         }
     }
 
-    function handleIdentifier(node, parent)
-    {
+    //
+    // Managing Block-Scope Model, push/pop & getting variables.
+    //
+
+    /**
+     * This function is called by the Traverser. 
+     * 
+     * @param  {BlockStatement|Program} blockNode
+     * @return {void}
+     */
+    function pushBlockScope(blockNode){
+        let index = scopeChain.indexOf(blockNode);
+        if(index !== -1){
+            throw Error("pushBlockScope() block-scope are already in scope chain");
+        }
+        scopeChain.push(blockNode);
+        currentScope = blockNode;
+    }
+
+    /**
+     * This function is called by the Traverser.
+     * 
+     * @param  {BlockStatement|Program} blockNode 
+     * @return {BlockStatement|Program}
+     */
+    function popBlockScope(blockNode){
+        let scope = scopeChain.pop();
+
+        if(scope !== blockNode){
+            throw Error("popBlockScope() where not equal to the asserted scope object");
+        }
+        let len = scopeChain.length;
+        if(len > 0){
+            currentScope = scopeChain[len -1];
+        }else{
+            currentScope = null; 
+        }
+
+        return scope;
+    }
+
+    function scopeChainHasVariable(name){
+
+    }
+
+    function scopeChainHasTypedVariable(name){
+
+    }
+
+    /**
+     * Returns the scope variable definition from the first scope closest relative to the current scope.
+     * 
+     * @param  {String} name 
+     * @return {[type]}      [description]
+     */
+    function scopeChainGetVariable(name){
+        if(!currentScope){
+            return null;
+        }
+        let cScope, scopeVars = currentScope._scopeVars;
+
+        if(scopeVars.hasOwnProperty(name)){
+            return scopeVars[name];
+        }
+
+        let startIdx = scopeChain.length;
+
+        if(startIdx == 1){
+            return null;
+        }
+
+        startIdx = startIdx - 2;
+        // backwards loop over scope chain as the begining of the array is top-level and the end its local.
+        for(let i = startIdx; i >= 0;i--){
+            cScope = scopeChain[i];
+            scopeVars = cScope._scopeVars;
+            if(scopeVars.hasOwnProperty(name)){
+                return scopeVars[name];
+            }
+        }
+
+        return null;
+    }
+
+    //
+    // Begining of Raweden Modification
+    //
+
+    /**
+     * Validates any Usage of the global property $oj_oj/$obj_ns
+     * 
+     * @param  {[type]}  node   [description]
+     * @param  {[type]}  parent [description]
+     * @return {Boolean}        [description]
+     */
+    function isValidRuntimeUsage(node, parent){
+        let oj_parent = parent.oj_parent;
+        if(oj_parent.type !== "CallExpression"){
+            return false;
+        }
+
+        let propertyName = parent.property.name;
+        return NSRuntimeModel.hasCallable(propertyName);
+    }
+
+    function getInlineBlockStatement(node){
+        let chain = [];
+        let cnode = node;
+        while(cnode !== null){
+            if(cnode.type == Syntax.BlockStatement){
+                return cnode;
+            }
+            chain.push(cnode);
+            cnode = cnode.oj_parent ? cnode.oj_parent : null;
+            // prevents circular references!
+            if(chain.indexOf(cnode) !== -1){
+                return null;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * @deprecated Moved to Builder.js
+     * @todo We need to take into account that the block scope may have arguments that may be used in the same way as variables.
+     * 
+     * @param  {}
+     * @return {[type]}
+     */
+    function parseBlockScope(blockNode){
+        console.log("-- start: handleIdentifier::parseBlockScope() --");
+        
+        let blockScope = {};
+        let i, len, node, nodes, type;
+
+        // handle block statements which parent are a NilScript method. arguments are part of the closure scoped variables.
+        if(blockNode.oj_parent.type == "OJMethodDefinition" && blockNode.oj_parent.methodSelectors){
+            let method = blockNode.oj_parent;
+            nodes = method.methodSelectors;
+            len = nodes.length;
+            for(i = 0;i < len;i++){
+                node = nodes[i];
+                if(!node.variableName){
+                    continue;
+                }
+                let scopeVar = {};
+                scopeVar.kind = "arg"; // var/let/const/arg
+                scopeVar.name = node.variableName.name;
+
+                scopeVar.typed = node.methodType && node.methodType.type == Syntax.OJParameterType ? node.methodType.value : null;
+
+                scopeVar.declaration = node;
+
+                // we need to assert that scopeVar.name is a string value.
+                if(typeof scopeVar.name !== "string"){
+                    continue;
+                }
+
+                blockScope[scopeVar.name] = scopeVar;
+            }
+            // Working with Parameters/Arguments.
+            // blockNode.oj_parent.methodSelectors[i].methodType.value  === gets the annotateted type for the argument.
+            // blockNode.oj_parent.methodSelectors[i].variableName.name === gets the defined argument name.
+            // 
+            // working with selectors:
+            // blockNode.oj_parent.methodSelectors[i].name              === gets the name segment to the argument point.
+            // blockNode.oj_parent.methodSelectors[i].returnType        === gets the return type declared by the selector.
+            // blockNode.oj_parent.methodSelectors[i].selectorType      === selector type either "-" or "+"
+        }else if(blockNode.oj_parent == Syntax.FunctionDeclaration){
+            // oj_parent: FunctionDeclaration
+            //     annotation: null
+            //     async: false
+            //     body: BlockStatement {type: "BlockStatement", body: Array(13), loc: {…}, oj_parent: FunctionDeclaration}
+            //     expression: false
+            //     generator: false
+            //     id: Identifier {type: "Identifier", name: "main", annotation: null, loc: {…}, oj_parent: FunctionDeclaration}
+            //     loc: {start: {…}, end: {…}}
+            //     oj_parent: Script {type: "Program", body: Array(6), sourceType: "script", loc: {…}}
+            //     params: Array(1)
+            //         0: Identifier || OJIdentifierWithAnnotation
+            //             annotation: null
+            //             loc: {start: {…}, end: {…}}
+            //             name: "argv"
+            //             oj_parent: FunctionDeclaration {type: "FunctionDeclaration", id: Identifier, params: Array(1), body: BlockStatement, generator: false, …}
+            //             type: "Identifier" // (in either case always this type)
+            //             __proto__: Object
+            //         length: 1
+            //         __proto__: Array(0)
+            //     type: "FunctionDeclaration"
+            //
+            //  Annotation on Parameter
+            //     annotation: OJTypeAnnotation
+            //          loc: {start: {…}, end: {…}}
+            //          oj_parent: OJIdentifierWithAnnotation {type: "Identifier", name: "argv", annotation: OJTypeAnnotation, loc: {…}, oj_parent: FunctionDeclaration}
+            //          optional: false
+            //          type: "OJTypeAnnotation"
+            //          value: "Object"
+        }
+
+        // lets get inside the scope and get all defined var(s) and let(s)
+        nodes = blockNode.body;
+        len = nodes.length;
+
+        for(i = 0;i < len;i++){
+            node = nodes[i];
+            if(node.type !== Syntax.VariableDeclaration){
+                continue;
+            }
+            let varKind = node.kind;
+            let vard, vard_id, vars = node.declarations;
+            let noVars = vars.length;
+
+            for(let v = 0;v < noVars;v++){
+                vard = vars[v]; // variable declarion object (esprisma: VariableDeclarator)
+                if(!vard.id){
+                    continue;
+                }
+
+                vard_id = vard.id;
+                let varName = vard_id.name;
+
+                //let varId = typeof vard.id && vard.id.type == ";
+                let scopeVar = {};
+                scopeVar.kind = varKind; // var/let/const/arg
+                scopeVar.name = varName;
+                scopeVar.typed = vard_id.annotation && vard_id.annotation.type == "OJTypeAnnotation" ? vard_id.annotation.value : null;
+
+                scopeVar.declaration = vard;
+
+                blockScope[varName] = scopeVar;
+            }
+
+        }
+        blockNode.scope = blockScope;
+    }
+
+    /**
+     * Walks the node chain by parent upwards until it finds the context in which `self` are defined.
+     * 
+     * @param  {}
+     * @return {OJClassImplementation}
+     */
+    function getSelfContext(node){
+        let chain = [];
+        let cnode = node;
+        while(cnode !== null){
+            if(cnode.type == "OJClassImplementation"){
+                return cnode;
+            }
+            chain.push(cnode);
+            cnode = cnode.oj_parent ? cnode.oj_parent : null;
+            // prevents circular references!
+            if(chain.indexOf(cnode) !== -1){
+                return null;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * Helper function which searches within the given class and sub-classing chain to retrive the 
+     * data model of a defined property (defined within source code by the @property directive). 
+     * 
+     * @param  {String} className    The name of the class in which to search for the property.
+     * @param  {String} propertyName The property name for which to search for.
+     * @return {OJProperty}          A OJProperty object instance or null if the property where not found.
+     */
+    function getPropertyInClassOrSubclass(className, propertyName){
+        if(typeof model._builtClassModelCache == "undefined"){
+            // we need to build a resolved class model cache. Where we cache:
+            // - a list of all properties defined with @property by the class and all its subclasses.
+            // - a list of all ivars defined by the class and all its subclasses.
+            // 
+            // It comes at the cost of a little bit more allocated memory when compiling.
+            // but would however speed up the compilation process as we dont need to loop
+            // over each subclass each time we want to know if a class defines a given
+            // property.
+        }
+
+        let classModels = model.classes;
+        let classModel = classModels.hasOwnProperty(className) ? classModels[className] : null;
+        let classChain = []; // pushes each name in here to prevent to get stuck in a infintive loop.
+        let propertyMap, cName;
+
+        // before we enter the while loop, lets add the first class to the chain.
+        classChain.push(className);
+
+        while(classModel !== null){
+
+            propertyMap = classModel._propertyMap
+
+            if(propertyMap.hasOwnProperty(propertyName) && typeof propertyMap[propertyName] == "object"){
+                return propertyMap[propertyName];
+            }
+
+            // setting class model for next loop, if we have a super class otherwise we return null.
+            if(typeof classModel.superclassName == "string" && classChain.indexOf(classModel.superclassName) == -1){
+                cName = classModel.superclassName;
+                classModel = classModels.hasOwnProperty(cName) ? classModels[cName] : null;
+                classChain.push(cName);
+            }else{
+                return null;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * Helper function which retrives the data-model for a instance-variable (iVar) declared by a class, it starts by
+     * looking at the class referenced by `className` and if its defined by that class its looks within classes
+     * declared in the sub-classing chain.
+     * 
+     * @param  {String} className The name of the class in which to search for the ivar name.
+     * @param  {String} ivarName  A string value that specifies the instance-variable name to search for.
+     * @return {OJIvar}           A OJIvar object instance or null if the property where not found.
+     */
+    function getIvarInClassOrSubclass(className, ivarName){
+        // use built class model cache as mentioned above.
+
+        let classModels = model.classes;
+        let classModel = classModels.hasOwnProperty(className) ? classModels[className] : null;
+        let classChain = []; // pushes each name in here to prevent to get stuck in a infintive loop.
+        let propertyMap, cName;
+
+        // before we enter the while loop, lets add the first class to the chain.
+        classChain.push(className);
+
+        while(classModel !== null){
+
+            propertyMap = classModel._ivarMap
+
+            if(propertyMap.hasOwnProperty(ivarName) && typeof propertyMap[ivarName] == "object"){
+                return propertyMap[ivarName];
+            }
+
+            // setting class model for next loop, if we have a super class otherwise we return null.
+            if(typeof classModel.superclassName == "string" && classChain.indexOf(classModel.superclassName) == -1){
+                cName = classModel.superclassName;
+                classModel = classModels.hasOwnProperty(cName) ? classModels[cName] : null;
+                classChain.push(cName);
+            }else{
+                return null;
+            }
+        }
+
+        return null;
+    }
+
+    const nativeTypes = ["String", "Boolean", "Number", "RegExp", "Array", "Object", "Function"];
+
+    function typeIsNativeJS(typeName){
+        return nativeTypes.indexOf(typeName) !== -1;
+    }
+
+    /**
+     * Translates a Selector name to its runtime name.
+     * @param  {[type]} selName [description]
+     * @return {[type]}              [description]
+     */
+    function translateSelectorName(selName){
+        if(selName.indexOf(":")){
+            selName = selName.replace(":", "_");
+        }
+        return "$oj_f_" + selName;
+    }
+
+    function AssertSameLine(node1, node2){
+        if(node1.loc.end.line !== node1.loc.start.line){
+            throw new Error("Expected assignment to be on the same line! At line: " + node1.loc.end.line);
+        }
+    }
+
+    /**
+     * Replaces the content inbetween Two Nodes Used within the modification of the generator to replace 
+     * the operator which is not part of the AST.
+     * 
+     * @param  {[type]} leftNode    [description]
+     * @param  {[type]} rightNode   [description]
+     * @param  {[type]} replacement [description]
+     * @return {[type]}             [description]
+     */
+    function replaceBetween(leftNode, rightNode, replacement){
+        let lineNo = leftNode.loc.end.line;
+        let fromColumn = leftNode.loc.end.column;
+        let toColumn = rightNode.loc.start.column;
+        AssertSameLine(leftNode, rightNode);
+
+        if(typeof replacement != "string"){
+            replacement = "";
+        }
+
+        modifier._addReplacement(lineNo, fromColumn, toColumn, replacement);
+    }
+
+    /**
+     * Handles MemberExpression generation based on block scoped defined variables and annoations.
+     *
+     * Handles compile time generation of:
+     * - translation to accessors in code when using dot notation on @property (requires annotation when not in self context)
+     * - translation of instance-variables to the compile-time names.          (requires annotation when not in self context)
+     *
+     * @todo  Walk to parent blockstatement if not found in the first, and so on.
+     * @todo  Should check for @property and ivar on subclasses if not found in the first. should walk the whole subclass chain if needed.
+     * 
+     * @param  {[type]} node   [description]
+     * @param  {[type]} parent [description]
+     * @return {[type]}        [description]
+     */
+    function handleMemberExprAdvanced(node, parent){
+        // we need to know the block-scope chain to be able to do some assertments.
+        //let blockNode = getInlineBlockStatement(node.oj_parent);
+        //if(!blockNode){
+        //    return;
+        //}
+        //if(typeof blockNode.scope === "undefined"){
+        //    parseBlockScope(blockNode);
+        //}
+
+        let expr = node.oj_parent; // no need to assert, as we would not passed the if above otherwise.
+        let objectId = expr.object;
+        let propertyId = expr.property;
+        let propertyName = propertyId.name;
+
+        if(objectId.name == "self"){
+            // Handles member expression where the object in the expression is self.
+            let selfCtxNode = getSelfContext(node);
+            let selfClassName = selfCtxNode.id.name;
+            let action = "none";
+            let fn;
+            let replacement;
+
+            // Case: Getting variable declarion and initial state is MemberExpression, then its a getter operation.
+            if(expr.oj_parent.type == Syntax.VariableDeclarator && expr.oj_parent.init === expr){
+                action = "get";
+            }
+
+            // Case: Getting - The member expression is at the right side in the assignment expression.
+            if(expr.oj_parent.type == Syntax.AssignmentExpression && expr.oj_parent.right === expr){
+                action = "get";
+            }
+
+            // Case: Setting - The member expression is at the left side in the assignment expression.
+            if(expr.oj_parent.type == Syntax.AssignmentExpression && expr.oj_parent.left === expr){
+                action = "set";
+            }
+            
+  
+            //console.log("Getting self in context of " + selfCtxNode.id.name);
+
+            let selfClass = model.classes.hasOwnProperty(selfClassName) ? model.classes[selfClassName] : null;
+            // first we check if there is a @property defined in the model for the class.
+            //let propertyDef = selfClass._propertyMap && selfClass._propertyMap.hasOwnProperty(propertyId.name) ? selfClass._propertyMap[propertyId.name] : null;
+            let propertyDef = getPropertyInClassOrSubclass(selfClassName, propertyName);
+            if(propertyDef){
+                // handles the case were the dot notation is pointing to a defined setter/getter trough @property directive.
+                
+                //console.log(selfClassName + " has @property for: " + propertyId.name + " should replace member-expression with " + action + " call");
+
+
+                // Case: Getting variable declarion and initial state is MemberExpression, then its a getter operation.
+                if(expr.oj_parent.type == Syntax.VariableDeclarator && expr.oj_parent.init === expr){
+                    replacement = translateSelectorName(propertyDef.getter) + "()";
+                    modifier.select(propertyId).replace(replacement);
+                }
+
+                // Case: Getting - The member expression is at the right side in the assignment expression.
+                if(expr.oj_parent.type == Syntax.AssignmentExpression && expr.oj_parent.right === expr){
+                    replacement = translateSelectorName(propertyDef.getter) + "()";
+                    modifier.select(propertyId).replace(replacement);
+                }
+
+                // Case: Setting - The member expression is at the left side in the assignment expression.
+                if(expr.oj_parent.type == Syntax.AssignmentExpression && expr.oj_parent.left === expr){
+                    if(expr.oj_parent.right.type == Syntax.Identifier){
+                        // getting the genrated name for property setter.
+                        replacement = translateSelectorName(propertyDef.setter);
+                        // replaces the operator.
+                        replaceBetween(expr.oj_parent.left, expr.oj_parent.right, "");
+
+                        modifier.select(propertyId).replace(replacement);
+                        replacement = "(" + expr.oj_parent.right.name + ")";
+                        modifier.select(expr.oj_parent.right).replace(replacement);
+                    }else if(expr.oj_parent.right.type == Syntax.Literal){
+                        //console.log("handleMemberExprAdvanced() handle Literal assignment to setter");
+                        // litieral can be anyting from String, Number, Boolean... to complex as Array and Object notations.
+                        // support String, Number and Boolean inline.
+                        // support Object & Array by asigning it to a temporary variable.
+                        //console.log(expr.oj_parent.right);
+                    }else{
+                        // create a temporary variable to to hold the evaluated value of the right parameter 
+                        // and then call the setter with that value..
+                        //console.log("------ERROR! ADVANCED SETTER COULD NOT BE GENERATED!------")
+                    }
+                }
+
+                
+            }
+            // secondly we check if there is a ivar declared with that name.
+            //let ivarDef = selfClass._ivarMap && selfClass._ivarMap.hasOwnProperty(propertyId.name) ? selfClass._ivarMap[propertyId.name] : null;
+            let ivarDef = getIvarInClassOrSubclass(selfClassName, propertyName);
+            if(ivarDef){
+                replacement = symbolTyper.getSymbolForClassNameAndIvarName(ivarDef.className, propertyId.name);
+                modifier.select(propertyId).replace(replacement);
+                //console.log("handleMemberExprAdvanced() " + selfClassName + " has instance-variable for: " + propertyId.name);
+            }
+            return;
+        }
+
+        let varInScope = scopeChainGetVariable(objectId.name);
+
+        //if(blockNode.scope.hasOwnProperty(objectId.name) && blockNode.scope[objectId.name].typed !== null){
+        if(varInScope && varInScope.typed !== null){
+            // Handles member-expression that finds typed matches in the local block scope. For example:
+            // 
+            // let obj:MyClass ...code that returns the initital value...
+            // obj.myProperty = aValue;
+            //
+            // Where this block closure looks at the defined class(es) and determine if we need to alter the code to point to a accessor,
+            // or that we need to point to a instance-variable within object.
+            //let objClassName = blockNode.scope[objectId.name].typed;
+            let objClassName = varInScope.typed;
+            // if the type declared as native JavaScript type we should not continue.
+            let isNative = typeIsNativeJS(objClassName);
+            if(isNative){
+                return false;
+            }
+
+            let objClass = model.classes.hasOwnProperty(objClassName) ? model.classes[objClassName] : null;
+            let fn, replacement;
+            //console.log("handleMemberExprAdvanced() " + objectId.name + " where found in the block-scope where it annotated type is set to: " + objClassName);
+            if(!objClass){
+                return false;
+            }
+            // first of we check for @property defined on the instance of class, and subclasses.
+            //let propertyDef = objClass._propertyMap && objClass._propertyMap.hasOwnProperty(propertyId.name) ? objClass._propertyMap[propertyId.name] : null;
+            let propertyDef = getPropertyInClassOrSubclass(objClassName, propertyName);
+            if(propertyDef){
+                // handles the case were the dot notation is pointing to a defined setter/getter trough @property directive.
+                //console.log("handleMemberExprAdvanced() " + objClassName + " has @property for: " + propertyId.name + " should replace member-expression with call");
+
+                var expCtx = expr.oj_parent;
+
+                if(expCtx.type == Syntax.VariableDeclarator || (expCtx.type == Syntax.AssignmentExpression && expCtx.right == expr)){
+                    // first  condition: A type of Assignment Expression where we set the initial value of a variable.
+                    // second condition: Assignment to another variable, therefor we should replace with getter accessor.
+                    
+                    // Replacing the property with the getter accessor. Simple work!
+                    replacement = translateSelectorName(propertyDef.getter) + "()";
+                    modifier.select(propertyId).replace(replacement);            
+
+                }else if(expCtx.type == Syntax.AssignmentExpression && expCtx.left == expr){
+                    // condition: We are on the left side in a Assignment Expression and should replace property with setter method call.
+                    
+                    if(expCtx.right.type == Syntax.Literal){
+                        // expCtx.right.raw <-- the raw string value as within script.
+                        // expCtx.right.value <-- evaluated to JavaScript type.
+                        replacement = translateSelectorName(propertyDef.setter);
+
+                        // removes the operator in the output.
+                        replaceBetween(expCtx.left, expCtx.right, "");
+
+                        modifier.select(propertyId).replace(replacement);
+                        replacement = "(" + expCtx.right.raw + ")";
+                        modifier.select(expCtx.right).replace(replacement);
+                    }else if(expCtx.right.type == Syntax.Identifier){
+                        replacement = translateSelectorName(propertyDef.setter);
+                        
+                        // removes the operator in the output.
+                        replaceBetween(expCtx.left, expCtx.right, "");
+
+                        modifier.select(propertyId).replace(replacement);
+                        modifier.before(expCtx.right).insert("(");
+                        //replacement = "(" + expCtx.right.name + ")";
+                        modifier.after(expCtx.right).insert(")");                        
+                    }else{
+                        // otherwise should assign it to a temporary variable and then use it for the assignment.
+                        // currently this just inlines it with paratences around..
+                        replacement = translateSelectorName(propertyDef.setter);
+                        
+                        // removes the operator in the output.
+                        replaceBetween(expCtx.left, expCtx.right, "");
+
+                        modifier.select(propertyId).replace(replacement);
+                        modifier.before(expCtx.right).insert("(");
+                        //replacement = "(" + expCtx.right.name + ")";
+                        modifier.after(expCtx.right).insert(")");                          
+                    }
+                }
+
+            }
+            // secondly we check if there is a ivar declared with that name.
+            //let ivarDef = objClass._ivarMap && objClass._ivarMap.hasOwnProperty(propertyId.name) ? objClass._ivarMap[propertyId.name] : null;
+            let ivarDef = getIvarInClassOrSubclass(objClassName, propertyName);
+            if(ivarDef){
+                //console.log("handleMemberExprAdvanced() " + objClassName + " has instance-variable for: " + propertyId.name);
+                replacement = symbolTyper.getSymbolForClassNameAndIvarName(ivarDef.className, propertyId.name);
+                modifier.select(propertyId).replace(replacement);
+            }
+        }
+
+        //console.log("handleIdentifier() parent is MemberExpression && grandparent is VariableDeclarator (logging: node.oj_parent.oj_parent)");
+        //console.log(node.oj_parent.oj_parent);
+        //console.log("-- start: handleIdentifier::BlockStatement{} --");
+
+        //console.log(blockNode);
+        //console.log("-- end: handleIdentifier::BlockStatement{} --");
+    }
+
+    //
+    // End of Raweden Modification.
+    //
+
+    /**
+     * @param  {[type]}
+     * @param  {[type]}
+     * @return {void}
+     */
+    function handleIdentifier(node, parent){
+
         let name   = node.name;
         let isSelf = (name == "self");
+
+        /*
+        // Temporary making some test for implementing at compile-time code generation for setter and getters.
+        // @todo: we need to handle the case for computed MemberExpressions!
+        if(name == "myStringProperty" || name == "m_MyStringProperty" || name == "myAnyProperty" || name == "m_propInBase"){
+            // before we run this code without the predefined name check of "myStringProperty" we need to know if the identifier could be a @property access.
+            if(typeof node.oj_parent !== "undefined" && typeof node.oj_parent.oj_parent !== "undefined"){
+                // parent is a member-expression and the parent of that is a Assignment Expression, we can be on both sides setter/getter
+                if(node.oj_parent.type == Syntax.MemberExpression && (node.oj_parent.oj_parent.type == Syntax.AssignmentExpression || node.oj_parent.oj_parent.type == Syntax.VariableDeclarator)){
+
+                    handleMemberExprAdvanced(node, parent);
+                }
+            }
+        }
+        */
+
+        if(parent.type == Syntax.MemberExpression && parent.property == node){
+            //console.log("handleIdentifier() --> we are in a MemberExpression where the current node is the property identifier");
+            handleMemberExprAdvanced(node, parent);
+        }
 
         if (name[0] === "$"){
             // determines if the access to $ (dollar) is accessing the underlaying objc like callable like $oj_oj.msgSend() then we dont throw a error here.
@@ -764,6 +1422,14 @@ generate()
             handleOJPredefinedMacro(node);
             return;
         }
+
+        //
+        // @todo: The code below would be better of replace with a stack based variable scope driven approach.
+        //        1. We need to check if the identifier is in local scope, then it should not be replace with the $oj_oj._g replacement.
+        //        2. Type check in local scope to allow access to class instance variables. In objc this is done with -> but we can use a simple dot (.)
+        //        3. Access to @property defined getters and setters needs to be generated inside impl and when type is annotated.
+        //
+        //
 
         if (!isIdentifierTransformable(node)) return;
 
@@ -796,7 +1462,7 @@ generate()
                 modifier.select(node).replace(replacement);
                 return;
 
-            } else {
+            }else{
                 if (name[0] == "_" && optionWarnUnknownIvars && (name.length > 1)) {
                     warnings.push(Utils.makeError(OJWarning.UndeclaredInstanceVariable, "Use of undeclared instance variable " + node.name, node));
                 }
@@ -824,6 +1490,7 @@ generate()
             }
         }
     }
+
 
     function handleVariableDeclaration(node, parent)
     {
@@ -1395,6 +2062,10 @@ generate()
             if (optionWarnDebugger) {
                 warnings.push(Utils.makeError(OJWarning.UseOfDebugger, "Use of debugger statement", node));
             }
+        }else if(type === Syntax.BlockStatement){
+            pushBlockScope(node);
+        }else if(type == Syntax.Program){
+            pushBlockScope(node);
         }
 
     }, function(node, parent) {
@@ -1420,9 +2091,14 @@ generate()
 
         } else if (type === Syntax.FunctionDeclaration || type === Syntax.FunctionExpression || type == Syntax.ArrowFunctionExpression) {
             finishScope(scope);
+        }else if(type === Syntax.BlockStatement){
+            popBlockScope(node);
+        }else if(type == Syntax.Program){
+            popBlockScope(node);
         }
 
         if (scope.node === node) {
+            //console.log(scope);
             scope = scope.previous;
         }
     });